backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
qdrant-client==1.9.1
httpx==0.25.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
openai==1.3.6
tiktoken==0.5.2
markdown==3.5.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-slugify==8.0.1
asyncpg==0.29.0
alembic==1.13.1
beautifulsoup4==4.12.2
scikit-learn==1.3.2
google-generativeai==0.8.6